+++ CHANGES.md
@@ -1,3 +1,9 @@
+# Unreleased (v1.1.0)
+
+* requires Thredded 1.0.1+
+* use allowlist rather than whitelist
+
+
 # v1.0.0
 
 Support ruby 3.0